feat(game): split game rule
