--- game/__init__.py
@@ -1,1 +1,0 @@
-from .game import Game
