+++ resources/images/empty_point.png
